Weapon - PsychoAxe 구현 중 - 이전 랜더링이 되지 않는 문제는 lifeTime 설정이 잘못된 것인 원이이었음을 확인 후 현재 수정 - 현재 처음 몇 Axe의 크기가 비정상적인 크기로 생성되는 것을 확인, 또한 Axe의 이동 및 회전이 정상적이지 않음을 확인 수정 예

+++ DX_MyProject/Object/Skill/Weapon/Weapon.h
@@ -29,7 +29,7 @@ protected:
 	vector<int> hitLimit_table;
 	vector<float> delay_table;
 	vector<int> ricochet_table;
-	vector<int> projLifetime_table;
+	vector<float> projLifetime_table;
 
 	// 무기에서 사용하는 projectile
 	vector<Projectile*> projectiles;
